Ensure errors are caught
This is some weird wonkery, but it allows the main entrypoint to catch
rejections (thus avoiding UnhandledRejection exceptions);
while at the same time allowing the un-caught promise to be exported in
such a way that other modules may require it directly.

The error utility is only necessary because core.setFailed doesn't
accept Errors as an argument.